requirements.txt
blinker           >=1.7.0
certifi           >=2023.11.17
charset-normalizer>=3.3.2
click             >=8.1.7
flask             >=3.0.0
gunicorn          >=21.2.0
idna              >=3.5
importlib-metadata>=6.8.0
imutils           >=0.5.4
itsdangerous      >=2.1.2
Jinja2            >=3.1.2
MarkupSafe        >=2.1.3
numpy             >=1.26.2
opencv-python     >=4.5.3.56
packaging         >=23.2
pip               >=20.3.4
pkg-resources     >=0.0.0
pushbullet.py     >=0.12.0
python-magic      >=0.4.27
requests          >=2.31.0
setuptools        >=44.1.1
urllib3           >=2.1.0
v4l2-python3      >=0.3.2
websocket-client  >=1.6.4
werkzeug          >=3.0.1
wheel             >=0.41.3
zipp              >=3.17.0